assignColors

Assign to each value in x a color according to the choices of breaks and col.

Usage

assignColors(x, breaks = NULL, col = heat.colors, na.col = "white")

Arguments

x

numeric or non-numeric vector

breaks

vector with breaks

col

vector with colors or color function

na.col

color for NA or out-of-range values

Value

vector of color with the same length as x with the attributes breaks the breaks used, col the color coding and na.col

the color for NA and out-of-range entries

Details

Depending if x is a numeric or non-numeric vector colors are assigned to each value.

In case of a numeric vector breaks can be

a number, giving the number of intervals covering the range of x,

a vector of two numbers, given the range to cover with 10 intervals, or

a vector with more than two numbers, specify the interval borders

In case of a non-numeric vector breaks must contain all values which are will get a color. If breaks is not given then a sensible default is choosen: in case of a numeric vector derived from pretty and otherwise all unique values/levels are used.

col can be either be a vector of colors or a function which generates via col(n) a set of n colors. The default is to use heat.colors.

Possible color functions in R packages can be found by vignette('plot.matrix').

Examples

## numeric vector
x <- runif(10)
assignColors(x)
#>  [1] "#FFFF80" "#FFAA00" "#FFFF00" "#FF0000" "#FFFF00" "#FFFF80" "#FF0000"
#>  [8] "#FFFF00" "#FFFF00" "#FF0000"
#> attr(,"breaks")
#> [1] 0.0 0.2 0.4 0.6 0.8 1.0
#> attr(,"col")
#> [1] "#FF0000" "#FF5500" "#FFAA00" "#FFFF00" "#FFFF80"
#> attr(,"na.col")
#> [1] "white"
# set breaks
assignColors(x, breaks=15)
#>  [1] "#FFFF2A" "#FFB900" "#FFD100" "#FF2E00" "#FFE800" "#FFFF2A" "#FF0000"
#>  [8] "#FFE800" "#FFD100" "#FF1700"
#> attr(,"breaks")
#>  [1] 0.00000000 0.06666667 0.13333333 0.20000000 0.26666667 0.33333333
#>  [7] 0.40000000 0.46666667 0.53333333 0.60000000 0.66666667 0.73333333
#> [13] 0.80000000 0.86666667 0.93333333 1.00000000
#> attr(,"col")
#>  [1] "#FF0000" "#FF1700" "#FF2E00" "#FF4600" "#FF5D00" "#FF7400" "#FF8B00"
#>  [8] "#FFA200" "#FFB900" "#FFD100" "#FFE800" "#FFFF00" "#FFFF2A" "#FFFF80"
#> [15] "#FFFFD5"
#> attr(,"na.col")
#> [1] "white"
assignColors(x, breaks=c(0,1))
#>  [1] "#FFFF40" "#FFB600" "#FFDB00" "#FF2400" "#FFFF00" "#FFFF40" "#FF0000"
#>  [8] "#FFFF00" "#FFDB00" "#FF0000"
#> attr(,"breaks")
#>  [1] 0.0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 1.0
#> attr(,"col")
#>  [1] "#FF0000" "#FF2400" "#FF4900" "#FF6D00" "#FF9200" "#FFB600" "#FFDB00"
#>  [8] "#FFFF00" "#FFFF40" "#FFFFBF"
#> attr(,"na.col")
#> [1] "white"
# set colors
assignColors(x, col=c("red", "green", "blue"))
#>  [1] "blue"  "green" "green" "red"   "blue"  "blue"  "red"   "blue"  "green"
#> [10] "red"  
#> attr(,"breaks")
#> [1] 0.0000000 0.3333333 0.6666667 1.0000000
#> attr(,"col")
#> [1] "red"   "green" "blue" 
#> attr(,"na.col")
#> [1] "white"
assignColors(x, col=topo.colors)
#>  [1] "#FFFF00" "#00E5FF" "#00FF4D" "#4C00FF" "#00FF4D" "#FFFF00" "#4C00FF"
#>  [8] "#00FF4D" "#00FF4D" "#4C00FF"
#> attr(,"breaks")
#> [1] 0.0 0.2 0.4 0.6 0.8 1.0
#> attr(,"col")
#> [1] "#4C00FF" "#004CFF" "#00E5FF" "#00FF4D" "#FFFF00"
#> attr(,"na.col")
#> [1] "white"
# NA and out-of-range
x[5] <- NA
assignColors(x, breaks=seq(0.5, 1, by=0.1), na.col="red")
#>  [1] "#FFFF00" "#FF0000" "#FF5500" "red"     "red"     "#FFFF00" "red"    
#>  [8] "#FFAA00" "#FF5500" "red"    
#> attr(,"breaks")
#> [1] 0.5 0.6 0.7 0.8 0.9 1.0
#> attr(,"col")
#> [1] "#FF0000" "#FF5500" "#FFAA00" "#FFFF00" "#FFFF80"
#> attr(,"na.col")
#> [1] "red"
## logical vector
l <- sample(c(NA, TRUE, FALSE), size=10, replace=TRUE)
assignColors(l)
#>  [1] "#FFFF00" "white"   "white"   "#FF0000" "white"   "white"   "#FF0000"
#>  [8] "#FFFF00" "#FF0000" "white"  
#> attr(,"breaks")
#> [1] "FALSE" "TRUE" 
#> attr(,"col")
#> [1] "#FF0000" "#FFFF00"
#> attr(,"na.col")
#> [1] "white"
assignColors(l, breaks=c("FALSE", "TRUE"), col=c("red", "blue"))
#>  [1] "blue"  "white" "white" "red"   "white" "white" "red"   "blue"  "red"  
#> [10] "white"
#> attr(,"breaks")
#> [1] "FALSE" "TRUE" 
#> attr(,"col")
#> [1] "red"  "blue"
#> attr(,"na.col")
#> [1] "white"
## character vector
t <- sample(letters, size=10, replace=TRUE)
assignColors(t)
#>  [1] "#FF4900" "#FF0000" "#FF9200" "#FF6D00" "#FFDB00" "#FFFF00" "#FFFFBF"
#>  [8] "#FF2400" "#FFB600" "#FFFF40"
#> attr(,"breaks")
#>  [1] "d" "i" "j" "k" "m" "n" "u" "v" "y" "z"
#> attr(,"col")
#>  [1] "#FF0000" "#FF2400" "#FF4900" "#FF6D00" "#FF9200" "#FFB600" "#FFDB00"
#>  [8] "#FFFF00" "#FFFF40" "#FFFFBF"
#> attr(,"na.col")
#> [1] "white"
assignColors(t, col=rainbow(5))
#>  [1] "#00FF66" "#FF0000" "#CC00FF" "#0066FF" "white"   "white"   "white"  
#>  [8] "#CCFF00" "white"   "white"  
#> attr(,"breaks")
#> [1] "d" "i" "j" "k" "m"
#> attr(,"col")
#> [1] "#FF0000" "#CCFF00" "#00FF66" "#0066FF" "#CC00FF"
#> attr(,"na.col")
#> [1] "white"